project_2/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
matplotlib
scikit-learn
seaborn